folders restructured for cat-be prep
