+++ src/pages/pass-finalisation/cat-b/pass-finalisation.cat-b.page.scss
@@ -1,4 +1,4 @@
-page-pass-finalisation-cat-b-page {
+pass-finalisation-cat-b-page {
   ion-content > div {
     background-color: map-get($colors, "mes-white");
   }
