+++ src/pages/post-debrief-holding/cat-b/post-debrief-holding.cat-b.page.html
@@ -1,7 +1,7 @@
 <ion-header>
   <practice-mode-banner *ngIf="isPracticeMode"></practice-mode-banner>
 </ion-header>
-<ion-content no-bounce id="post-debrief-holding-page">
+<ion-content no-bounce id="post-debrief-holding-cat-b-page">
   <div class="container">
     <img id='dvsa-logo' src='../../assets/imgs/back-to-office/logo.png' />
   </div>
