+++ src/pages/post-debrief-holding/cat-b/post-debrief-holding.cat-b.page.scss
@@ -1,4 +1,4 @@
-post-debrief-holding {
+post-debrief-holding-cat-b-page {
   div {
     background-color: map-get($colors-gds, gds-dark-green-2);
   }
